+++ api/src/main/java/org/allaymc/api/block/component/BlockStateDataComponent.java
@@ -1,6 +1,6 @@
-package org.allaymc.api.block.component.data;
+package org.allaymc.api.block.component;
 
-import org.allaymc.api.block.component.BlockComponent;
+import org.allaymc.api.block.data.BlockStateData;
 import org.allaymc.api.block.type.BlockState;
 
 /**
@@ -11,7 +11,6 @@ public interface BlockStateDataComponent extends BlockComponent {
      * Retrieves the {@link BlockStateData} associated with the specified {@link BlockState}.
      *
      * @param blockState the block state
-     *
      * @return the block state data for the given state
      */
     BlockStateData getBlockStateData(BlockState blockState);