+++ server/src/main/java/org/allaymc/server/registry/InternalRegistries.java
@@ -1,10 +1,10 @@
 package org.allaymc.server.registry;
 
-import org.allaymc.api.block.component.data.BlockStateData;
 import org.allaymc.api.block.data.BlockId;
+import org.allaymc.api.block.data.BlockStateData;
 import org.allaymc.api.block.property.type.BlockPropertyType;
+import org.allaymc.api.item.data.ItemData;
 import org.allaymc.api.item.data.ItemId;
-import org.allaymc.api.item.type.ItemData;
 import org.allaymc.api.registry.SimpleMappedRegistry;
 import org.allaymc.api.registry.SimpleRegistry;
 import org.allaymc.server.block.property.BlockPropertyProcessor;